feat: undo dropdown changes and added new `FloatingMenu` component
