+++ packages/ui/src/components/molecules/Dropdown/index.ts
@@ -1,5 +1,3 @@
 export * from './Dropdown';
-export * from './DropdownMenu';
 export * from './DropdownStyles';
-export * from './FloatingMenu';
 export * from './ListItemDropdown';
