+++ packages/ui/src/components/molecules/FloatingMenu/index.ts
@@ -1,0 +1,1 @@
+export * from './FloatingMenu';
